requirements.txt
markdown>=3.5
PyYAML>=6.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
